Add files via upload
Introduction:
As a part of development process of students of the college they are require to undertake trips to the various industries or anything else. This visit may be a part of their curricular or extra curricular activities to boost their enthusiasm within them.
So, this system is designed to manage data for the students who are willing to join the trip.